move changes into log package
