+++ src/main/java/log/Log4jLogger.java
@@ -1,4 +1,4 @@
-package misc;
+package log;
 
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.Logger;
